JBDS-4520 Attempt to fix unit tests
