+++ browser/services/data.js
@@ -29,10 +29,13 @@ class InstallerDataService {
     this.packageConf = packageConf;
 
     this.username = TokenStore.getUserName();
+    this.password = '';
     if (this.username) {
       let password = TokenStore.getItem('login', this.username);
       password.then((pass) => {
-        this.password = pass;
+        if(pass && pass !=='') {
+          this.password = pass;
+        }
       });
     }
 
@@ -198,20 +201,8 @@ class InstallerDataService {
   }
 
   localAppData() {
-    let appData = Platform.identify({
-      win32: ()=> {
-        let appDataPath = Platform.ENV.APPDATA;
-        return appDataPath ? path.join(appDataPath, '..', 'Local', 'RedHat', 'DevSuite') : this.tempDir();
-      }, darwin: ()=> {
-        let homePath = Platform.ENV.HOME;
-        return homePath ? path.join(homePath, 'Library', 'Application Support', 'RedHat', 'DevSuite') : this.tempDir();
-      }, default: ()=> {
-        return this.tempDir();
-      }
-    });
-    return path.resolve(appData);
+    return Platform.localAppData();
   }
-
 
   isDownloading() {
     return this.downloading;
